pyproject.toml
[tool.poetry]
name = "basana"
version = "1.2.1"
homepage = "https://github.com/gbeced/basana"
repository = "https://github.com/gbeced/basana"
documentation = "https://basana.readthedocs.io/en/latest/"
description = "A Python async and event driven framework for algorithmic trading, with a focus on crypto currencies."
authors = ["Gabriel Becedillas <gabriel.becedillas@gmail.com>"]
license = "Apache-2.0"
packages = [{include = "basana"}]

[tool.poetry.dependencies]
python = "^3.8.1"
aiohttp = {extras = ["speedups"], version = "^3.8.4"}
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
aioresponses = "^0.7.4"
flake8 = "^6.0.0"
mypy = "^1.0.1"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
talipp = "^1.9.1"
types-python-dateutil = "^2.8.19.8"
websockets = "^10.4"
sphinx = "^6.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

